Did the styling for the headers (the StackNavigation's headers).
. Note that I've decided to get rid of the headers in the GamePage and 
the HoorayPage all together. - Because the Back button was a bit fiddly 
to handle:
    .The Back button in this header is given the 'forfeitConfirmation' 
function only at the beginning of the render. And at that time, the 
gameOver == undefined (Because the state of the component hasn't loaded 
yet). This means that to the Back button's point of view, the gameOver 
== undefined forever, which causes a problem when trying to display the 
HoorayPage cause the Alert box will always show (because to the Back 
button's point of view, the gameOver == undefined).