Got the all the beads of the top rack to move up and down.
. I made a separate component called TopReackBead for the each bead in 
the top rack (instead of the previous normal Bead component).

. To get the bead to move up, you have to swipe up on the bead.
. To get the bead to move down, you have to swipe down on the bead.
. If the bead is already up, then it definitely cannot go up anymore 
(and a console log will tell you so).
. If the bead is already down, then it definitely cannot go down anymore 
(and a console log will tell you so).

. I left the console logs that tell whether the bead can move up or down 
or not, because swipping up and down with a mouse is difficult; so just 
for future reference, I'm leaving the console logs there.

. Also made a global constants file - and it currently has the bead 
height and width.