Finished configuring all the swiping up and down movements of the bottom rack.
. Everything in the bottom rack works as expected in real life (e.g. if 
the 4th bead is at a 'down' state, then swiping this 4th bead up and 
also push the other beads up if needed).

. Note that every time a bead(s) moves inside a single 
BottomRackColBeads component, this particular BottomRackColBeads 
component is the ONLY one that gets rerendered again - which is as 
expected. (And in turn of rerendering that single BottomRackColBeads 
component, it will also rerender its children (4 Bead components) - but 
I think thats ok)

. However, there's a rerendering issue:
    . Every time the abacus renders from toggling the abacus, all the 
BottomRackColBeads and all the TopRackBead comoponents get rendered 
TWICE (when I was more of expecting only rendering them once per 
toggle). - This will need to be fixed in the next commit.

. Note that I've still left all the console.logs cause they will still 
be needed in the near future.