Finished the 'Number of questions'screen.
Now this screen allows the user to pick 5, 10, 15 or 20 questions. After 
choosing which number of questions, the game will immediately start. 
Like just clicking the button with the number (5, 10, 15 or 20) on it 
will immeditely start the game; as in, there is no 'Next' button.

Note: I decided having default buttons of how many questions they want, 
instead of typing the exact number of questions, is better in terms of 
user experience and easier to develop in code too.