Totally cleaned up the game section AND fixed the hardware Back button issue.
The hardware Back button issue was that after you play the game again 
with the same configurations, the hardware Back button will not prompt 
to say whether you want to forfeit the game or not. 
It will do everything correctly the first time the game is played, but 
NOT do it correctly when playing the same game again.
So, that was fixed in this commit. Everything about the hardware Back 
button is working now; and I mean everything and everywhere too (no 
matter whether you are in the ready-made game or the custom-made game).

'Cleaned up the game section' means that we got rid of all the 
console.log.
So, now all the console.log of the whole app are gone.